+++ source/Examples/Resources/SqlScript/4-RunScriptUsingDisabledVariables.ps1
@@ -25,6 +25,7 @@ Configuration Example
             SetFilePath      = 'C:\DSCTemp\SQLScripts\Set-RunSQLScript.sql'
             TestFilePath     = 'C:\DSCTemp\SQLScripts\Test-RunSQLScript.sql'
             GetFilePath      = 'C:\DSCTemp\SQLScripts\Get-RunSQLScript.sql'
+            Variable         = @()
             DisableVariables = $true
         }
     }